server/requirements.txt
fastapi
uvicorn
langchain
langchain-community
openai
tqdm
qdrant-client
python-dotenv
motor